ai/schema_matcher/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
pydantic==2.10.3
requests==2.31.0
python-dotenv==1.0.0
